Factor out event handling code
